Partition datasets with respect to training data
Re-partition the positive and negative examples for the permuted hetnets
with respect to the original training data. The procedure now ensures
that all relations used as negative training examples are never
contained at all in the holdout set.

Previously the negative training examples may also show up as a possible
chemical-disease pair in the holdout set, but this data contamination
has been rectified.