Fix nothing being output to file
Move the dwpc function to inside the main function so that it can
access the writer lock variable. Previously it was out of scope
and therefore nothing was being written to the dwpc.tsv.bz2 file.

In addition, updated data access to be only compatible with
version 3 of py2neo.